+++ src/spring-cloud/HISTORY.md
@@ -3,9 +3,8 @@ Release History
 
 0.3.0
 -----
-* Add command "az spring-cloud update"
-* Support the tags update
-* Support to enable the distributed tracing when creating the service
+* Enable distributed tracing by default when creating the service
+* Add command "az spring-cloud update" to uppate tags and distributed tracing settings
 
 0.2.4
 -----